fix(kadena-cli): circular dependency
